Split UniformBuffer from Uniform.h
