+++ include/Engine/Renderer/Renderer.h
@@ -8,6 +8,7 @@
 #include "Engine/Renderer/Material/Material.h"
 #include "Engine/Renderer/Material/ReservedUniformBuffers.h"
 #include "Engine/Renderer/Material/Uniform.h"
+#include "Engine/Renderer/Material/UniformBuffer.h"
 #include "Engine/Renderer/Mesh/StaticMesh.h"
 #include "Engine/Renderer/Shader/ShaderProgram.h"
 #include "Engine/Renderer/Texture/Texture2D.h"
