+++ src/Engine/Renderer/Apis/Vulkan/Material/VulkanUniformBuffer.cpp
@@ -1,0 +1,10 @@
+#include "Engine/Utility/Core.h"
+
+#ifdef RENDERER_VULKAN
+
+#include "Engine/Renderer/Apis/Vulkan/Material/VulkanUniformBuffer.h"
+
+namespace gp1::renderer::vulkan
+{
+}
+#endif
